Moves DrakeVisualizerParams into dedicated header
This allows downstream code to make decisions about parameters without
dragging all of SceneGraph into their headers.
