+++ geometry/BUILD.bazel
@@ -93,6 +93,7 @@ drake_cc_library(
     srcs = ["drake_visualizer.cc"],
     hdrs = ["drake_visualizer.h"],
     deps = [
+        ":drake_visualizer_params",
         ":geometry_roles",
         ":geometry_version",
         ":scene_graph",
@@ -101,6 +102,15 @@ drake_cc_library(
         "//lcmtypes:viewer",
         "//systems/framework:context",
         "//systems/framework:leaf_system",
+    ],
+)
+
+drake_cc_library(
+    name = "drake_visualizer_params",
+    hdrs = ["drake_visualizer_params.h"],
+    deps = [
+        ":geometry_roles",
+        ":rgba",
     ],
 )
